% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{extract_data}
\alias{extract_data}
\alias{extract_incidents}
\alias{extract_residents}
\alias{extract_occupancy}
\alias{extract_reference_homes}
\alias{extract_tallies}
\alias{extract_timelines}
\title{Extract most recent flat files into R data objects}
\usage{
extract_incidents()

extract_residents()

extract_occupancy()

extract_reference_homes()

extract_tallies()

extract_timelines()

extract_data(reference_data = FALSE)
}
\arguments{
\item{reference_data}{a boolean indicating whether the reference data (one-off extracts)
should be extracted. The default is FALSE to save time.}
}
\value{
nothing
}
\description{
Saves the most recent data into directory `getOption("fshc_files")`
as files `"residents.rda"`, `"incident.rda"`, `"occupancy.rda"`, `"beds.rda"`
}
