% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{funnel_poisson_control_limits}
\alias{funnel_poisson_control_limits}
\title{Function factory for Poisson funnel plot control limits}
\usage{
funnel_poisson_control_limits(y, denom)
}
\arguments{
\item{y}{a numeric vector of observed counts}

\item{denom}{a numeric vector for the denominator (exposed to risk/person-days)}
}
\value{
a function taking the following arguments
\describe{
  \item{\code{x}}{value of the denominator}
  \item{\code{p}}{lower tail probability}
}
and returns the smallest integer y such that \eqn{Pr(X \ge y) = p}
}
\description{
Function factory for Poisson funnel plot control limits
}
\examples{
funnel_cl <- funnel_poisson_control_limits(10, 100)
funnel_cl(x = 100, .975)
}
