% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{get_monday_date}
\alias{get_monday_date}
\title{Get the week starting date}
\usage{
get_monday_date(date)
}
\arguments{
\item{date}{a `Date` vector}
}
\value{
a `Date` vector corresponding to the preceding Monday
}
\description{
Get the week starting date
}
