% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{get_time_series}
\alias{get_time_series}
\title{Build weekly time series table}
\usage{
get_time_series(timespan = get_time_span())
}
\arguments{
\item{timespan}{a list with a week start date and a week end date.
Default is \code{\link{get_time_span}()}.#'}
}
\value{
a data frame with variables `week_number`, `year`, `week_starting`,
and `week_ending`
}
\description{
Build weekly time series table
}
