% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{get_time_span}
\alias{get_time_span}
\title{Set timespan of measurement}
\usage{
get_time_span(data = incidents)
}
\arguments{
\item{data}{}
}
\value{
a list with two `Date` objects named `"start"` and `"end"`
}
\description{
Set timespan of measurement
}
